requirements.txt
configparser==3.7.4
dj-database-url==0.5.0
Django==2.2.2
django-appconf==1.0.3
django-ckeditor==5.7.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-debug-toolbar==2.0
django-formtools==2.1
django-hosts==3.0
django-js-asset==1.2.2
django-maintenancemode==0.11.3
django-model-utils==3.1.2
django-mptt==0.10.0
django-render-partial==0.4
IPy==1.0
mysql-connector-python==8.0.17
mysqlclient==1.4.4
Pillow==6.0.0
protobuf==3.9.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Werkzeug==0.15.4
